Use the new Cache API league/flysystem introduced in 1.0.0
This means we have to decorate the adapter with a CachedAdapter.
Note: This is not the optimal solution for this problem. We could
have used decorated services described here: http://symfony.com/doc/current/components/dependency_injection/advanced.html#decorating-services
But then again we would have to bumb the minimal version to 2.5.